requirements.txt
oandapyV20==0.6.3
pandas==2.2.0
numpy==1.26.4
python-dotenv==1.0.1
pyyaml==6.0.1
requests==2.31.0
pytest==8.1.1
pytest-cov==5.0.0
loguru==0.7.2
streamlit==1.33.0
python-telegram-bot==21.1.1
